test: Attendance 도메인 테스트 수정
