+++ src/test/java/com/nexters/momo/attendance/domain/AttendanceTest.java
@@ -1,6 +1,5 @@
 package com.nexters.momo.attendance.domain;
 
-import com.nexters.momo.attendance.exception.InvalidAttendanceCodeException;
 import com.nexters.momo.session.domain.Point;
 import com.nexters.momo.session.domain.Session;
 import org.junit.jupiter.api.DisplayName;
@@ -8,10 +7,11 @@ import org.junit.jupiter.api.Test;
 
 import java.time.LocalDateTime;
 
-import static com.nexters.momo.attendance.domain.AttendanceStatus.*;
+import static com.nexters.momo.attendance.domain.AttendanceStatus.ABSENT;
+import static com.nexters.momo.attendance.domain.AttendanceStatus.ATTENDANCE;
+import static com.nexters.momo.attendance.domain.AttendanceStatus.LATE;
 import static com.nexters.momo.session.domain.Session.createSession;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 /**
  * Attendance 엔티티 관련 테스트 클래스입니다.
@@ -23,8 +23,6 @@ class AttendanceTest {
 
     private static final long TEST_MEMBER_ID = 2L;
     private static final long TEST_GENERATION_ID = 22L;
-    Integer incorrectAttendanceCode;
-
 
     @DisplayName("출석 생성 테스트")
     @Test
@@ -38,34 +36,14 @@ class AttendanceTest {
         Session session = createSession("세션 키워드", 1, "세션 내용", sessionStartTime, sessionEndTime,
                 "서울시 강남구", "세부 주소", Point.of(20.002, 142.01), attendanceStartTime, attendanceEndTime, TEST_GENERATION_ID);
 
-        int attendanceCode = session.getAttendanceCode();
+        AttendanceStatus attendanceStatus = session.judgeAttendanceStatus();
 
         // when
-        Attendance attendance = Attendance.createAttendance(session, TEST_MEMBER_ID, attendanceCode);
+        Attendance attendance = Attendance.createAttendance(TEST_MEMBER_ID, session.getId(), attendanceStatus);
 
         // then
         assertThat(attendance.getStatus()).isEqualTo(ATTENDANCE);
 
-    }
-
-    @DisplayName("출석 코드 불일치 테스트")
-    @Test
-    void invalid_attendanceCode_test() {
-        // given
-        LocalDateTime sessionStartTime = LocalDateTime.now().plusMinutes(10);
-        LocalDateTime sessionEndTime = LocalDateTime.now().plusMinutes(100);
-        LocalDateTime attendanceStartTime = LocalDateTime.now().minusMinutes(10);
-        LocalDateTime attendanceEndTime = sessionEndTime;
-
-        Session session = createSession("세션 키워드", 1, "세션 내용", sessionStartTime, sessionEndTime,
-                "서울시 강남구", "세부 주소", Point.of(20.002, 142.01), attendanceStartTime, attendanceEndTime, TEST_GENERATION_ID);
-
-        incorrectAttendanceCode = session.getAttendanceCode() + 1;
-
-        // when, then
-        assertThatThrownBy(() ->
-                Attendance.createAttendance(session, TEST_MEMBER_ID, incorrectAttendanceCode)).
-                isInstanceOf(InvalidAttendanceCodeException.class);
     }
 
     @DisplayName("결석 테스트")
@@ -80,10 +58,8 @@ class AttendanceTest {
         Session session = createSession("세션 키워드", 1, "세션 내용", sessionStartTime, sessionEndTime,
                 "서울시 강남구", "세부 주소", Point.of(20.002, 142.01), attendanceStartTime, attendanceEndTime, TEST_GENERATION_ID);
 
-        int attendanceCode = session.getAttendanceCode();
-
         // when
-        Attendance attendance = Attendance.createAttendance(session, TEST_MEMBER_ID, attendanceCode);
+        Attendance attendance = Attendance.createAttendance(TEST_MEMBER_ID, session.getId(), session.judgeAttendanceStatus());
 
         // then
         assertThat(attendance.getStatus()).isEqualTo(ABSENT);
@@ -101,10 +77,8 @@ class AttendanceTest {
         Session session = createSession("세션 키워드", 1, "세션 내용", sessionStartTime, sessionEndTime,
                 "서울시 강남구", "세부 주소", Point.of(20.002, 142.01), attendanceStartTime, attendanceEndTime, TEST_GENERATION_ID);
 
-        int attendanceCode = session.getAttendanceCode();
-
         // when
-        Attendance attendance = Attendance.createAttendance(session, TEST_MEMBER_ID, attendanceCode);
+        Attendance attendance = Attendance.createAttendance(TEST_MEMBER_ID, session.getId(), session.judgeAttendanceStatus());
 
         // then
         assertThat(attendance.getStatus()).isEqualTo(LATE);
